requirements.txt
pydantic~=2.11.4
fastapi~=0.115.11
sqlalchemy~=2.0.41
sse-starlette~=2.3.3
requests~=2.32.3
langchain_deepseek~=0.1.2
openai~=1.74.0
langchain-core~=0.3.51
langchain~=0.3.20
python-dotenv~=1.0.1
uvicorn~=0.34.0